update main from original

+++ src/data_model.py
@@ -1,11 +1,9 @@
-# from __future__ import annotations
-
 def increment(x: int) -> int:
     return x + 1
 
 
 class Employee:
-    def __init__(self, name: str, age: int, salary: float) -> None:
+    def __init__(self, name: str, age: int, salary: int) -> None:
         self.name = name
         self.age = age
         self.salary = salary
